pyproject.toml
[project]
name = "color-mti-ch-exits"
version = "0.1.0"
description = ""
authors = [
    {name = "gvardovski",email = "gvardovski@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "vectorbt (>=0.28.2,<0.29.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "pandas-ta (>=0.4.71b0,<0.5.0)",
    "pyaml (>=25.7.0,<26.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false